remove test, rearrange files
